# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HATCHUP Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a new player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}/dashboard:
    get:
      operationId: getPlayerDashboard
      tags: [players]
      summary: Get player dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDashboard"

  # ── Hatchlings ────────────────────────────────────────────────────────────
  /hatchlings:
    get:
      operationId: listHatchlings
      tags: [hatchlings]
      summary: List hatchlings (optionally by owner)
      parameters:
        - name: playerId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of hatchlings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hatchling"
    post:
      operationId: createHatchling
      tags: [hatchlings]
      summary: Hatch a new creature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HatchlingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hatchling"
  /hatchlings/{id}:
    get:
      operationId: getHatchling
      tags: [hatchlings]
      summary: Get hatchling by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hatchling found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hatchling"
        "404":
          description: Not found
    patch:
      operationId: updateHatchling
      tags: [hatchlings]
      summary: Update hatchling (name, mood, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HatchlingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hatchling"
    delete:
      operationId: deleteHatchling
      tags: [hatchlings]
      summary: Release a hatchling
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /hatchlings/{id}/evolve:
    post:
      operationId: evolveHatchling
      tags: [hatchlings]
      summary: Trigger evolution for a hatchling
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvolveInput"
      responses:
        "200":
          description: Evolution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hatchling"
...
  /leaderboards/global:
    get:
      operationId: getGlobalLeaderboard
      tags: [leaderboards]
      summary: Get global player leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Global leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /leaderboards/by-mode:
    get:
      operationId: getModeLeaderboard
      tags: [leaderboards]
      summary: Get leaderboard for a specific game mode
      parameters:
        - name: mode
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Mode leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List available items in the shop/catalog
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/{id}/use:
    post:
      operationId: useItem
      tags: [items]
      summary: Use an item on a hatchling
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UseItemInput"
      responses:
        "200":
          description: Item used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hatchling"

  # ── Live Events ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, username, level, xp, coins, rank, totalWins, totalMatches, createdAt]
      properties:
        id:          { type: integer }
        username:    { type: string }
        displayName: { type: ["string", "null"] }
        avatarUrl:   { type: ["string", "null"] }
        level:       { type: integer }
        xp:          { type: integer }
        coins:       { type: integer }
        rank:        { type: string }
        rankScore:   { type: integer }
        totalWins:   { type: integer }
        totalMatches: { type: integer }
        clubId:      { type: ["integer", "null"] }
        createdAt:   { type: string }
    PlayerInput:
      type: object
      required: [username]
      properties:
        username:    { type: string, minLength: 2, maxLength: 20 }
        displayName: { type: string }
        avatarUrl:   { type: string }
    PlayerUpdate:
      type: object
      properties:
        displayName: { type: string }
        avatarUrl:   { type: string }
    PlayerDashboard:
      type: object
      required: [player, hatchlingCount, totalWins, recentCompetitions, activeEvents, topHatchling]
      properties:
        player:               { $ref: "#/components/schemas/Player" }
        hatchlingCount:       { type: integer }
        totalWins:            { type: integer }
        winRate:              { type: number }
        recentCompetitions:
          type: array
          items:
            $ref: "#/components/schemas/Competition"
        activeEvents:
          type: array
          items:
            $ref: "#/components/schemas/LiveEvent"
        topHatchling:         { $ref: "#/components/schemas/Hatchling" }

    # ── Hatchling ────────────────────────────────────────────────────────────
    Hatchling:
      type: object
      required: [id, playerId, name, species, evolutionStage, personality, mood, level, xp, happiness, hunger, energy, createdAt]
      properties:
        id:             { type: integer }
        playerId:       { type: integer }
        name:           { type: string }
        species:        { type: string }
        evolutionStage: { type: integer }
        evolutionType:  { type: ["string", "null"] }
        category:       { type: string }
        rarity:         { type: string }
        personality:    { type: string }
        mood:           { type: string }
        level:          { type: integer }
        xp:             { type: integer }
        happiness:      { type: integer }
        hunger:         { type: integer }
        energy:         { type: integer }
        abilityName:    { type: ["string", "null"] }
        abilityDesc:    { type: ["string", "null"] }
        imageUrl:       { type: ["string", "null"] }
        isShiny:        { type: boolean }
        isFusion:       { type: boolean }
        color:          { type: ["string", "null"] }
        createdAt:      { type: string }
    HatchlingInput:
      type: object
      required: [playerId, name]
      properties:
        playerId: { type: integer }
        name:     { type: string, minLength: 1, maxLength: 30 }
        species:  { type: string }
    HatchlingUpdate:
      type: object
      properties:
        name:      { type: string }
        mood:      { type: string }
        happiness: { type: integer }
        hunger:    { type: integer }
        energy:    { type: integer }
    EvolveInput:
      type: object
      required: [triggerId]
      properties:
        triggerId: { type: integer }

    # ── Evolution ────────────────────────────────────────────────────────────
...
    Competition:
      type: object
      required: [id, mode, status, playerId, hatchlingId, score, rank, createdAt]
      properties:
        id:          { type: integer }
        mode:        { type: string }
        status:      { type: string }
        playerId:    { type: integer }
        playerName:  { type: ["string", "null"] }
        hatchlingId: { type: integer }
        hatchlingName: { type: ["string", "null"] }
        score:       { type: integer }
        rank:        { type: ["integer", "null"] }
        duration:    { type: ["integer", "null"] }
        xpEarned:    { type: ["integer", "null"] }
        coinsEarned: { type: ["integer", "null"] }
        createdAt:   { type: string }
...
    LeaderboardEntry:
      type: object
      required: [position, playerId, username, rank, score, wins, hatchlingName]
      properties:
        position:      { type: integer }
        playerId:      { type: integer }
        username:      { type: string }
        displayName:   { type: ["string", "null"] }
        avatarUrl:     { type: ["string", "null"] }
        rank:          { type: string }
        score:         { type: integer }
        wins:          { type: integer }
        hatchlingName: { type: string }
        hatchlingCategory: { type: ["string", "null"] }
...
    Item:
      type: object
      required: [id, name, description, category, price, effectType, effectValue]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        category:    { type: string }
        price:       { type: integer }
        imageUrl:    { type: ["string", "null"] }
        effectType:  { type: string }
        effectValue: { type: integer }
        rarity:      { type: string }
        color:       { type: ["string", "null"] }
    UseItemInput:
      type: object
      required: [hatchlingId, playerId]
      properties:
        hatchlingId: { type: integer }
        playerId:    { type: integer }

    # ── Live Event ────────────────────────────────────────────────────────────
    LiveEvent:
      type: object
      required: [id, name, description, type, status, startsAt, endsAt, participants]
      properties:
        id:           { type: integer }
        name:         { type: string }
        description:  { type: string }
        type:         { type: string }
        status:       { type: string }
        startsAt:     { type: string }
        endsAt:       { type: string }
        participants: { type: integer }
        reward:       { type: ["string", "null"] }
        imageUrl:     { type: ["string", "null"] }
        color:        { type: ["string", "null"] }
        isFeatured:   { type: boolean }

    # ── Club ──────────────────────────────────────────────────────────────────
...