# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HATCHUP Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  name: string;
  description: string;
  category: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  effectType: string;
  effectValue: number;
  rarity?: string;
  /** @nullable */
  color?: string | null;
}