# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HATCHUP Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  level: number;
  xp: number;
  coins: number;
  rank: string;
  rankScore?: number;
  totalWins: number;
  totalMatches: number;
  /** @nullable */
  clubId?: number | null;
  createdAt: string;
}