# lib/api-zod/src/generated/types/useItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HATCHUP Game API
 * OpenAPI spec version: 0.1.0
 */

export interface UseItemInput {
  hatchlingId: number;
  playerId: number;
}